move AoMMState and AoMMParams into the Api file
